requirements.txt
google-genai
rich
python-dotenv
flask
flask-cors
gunicorn
